@@ -1,5 +1,5 @@
 {
-  "name": "perlur/php-fio-api",
+  "name": "ardani/php-fio-api",
   "type": "library",
   "description": "PHP Fio API implemention",
   "keywords": [
@@ -7,7 +7,7 @@
     "banking",
     "api"
   ],
-  "homepage": "https://github.com/perlur/php-fio-api",
+  "homepage": "https://github.com/ardani/php-fio-api",
   "license": "MIT",
   "authors": [
     {
@@ -23,6 +23,12 @@
       "role": "Software Developer"
     },
     {
+      "name": "Ardani Rohman",
+      "email": "master.ardani@gmail.com",
+      "homepage": "https://ardani.github.io",
+      "role": "Software Developer"
+    },
+    {
       "name": "Martin Hujer",
       "email": "mhujer@gmail.com",
       "homepage": "https://www.martinhujer.cz/",
@@ -30,17 +36,17 @@
     }
   ],
   "require": {
-    "php": ">=5.5",
+    "php": ">=7.1",
     "ext-curl": "*",
     "guzzlehttp/guzzle": "~6.1"
   },
   "require-dev": {
-    "phpunit/phpunit": "~4.7|~5.0",
-    "squizlabs/php_codesniffer": "~2.3",
-    "satooshi/php-coveralls": "^1.0"
+    "phpunit/phpunit": "~6.3",
+    "squizlabs/php_codesniffer": "~3.0",
+    "php-coveralls/php-coveralls": "^1.0"
   },
   "suggest": {
-    "kdyby/curl-ca-bundle": "Provides regularly updated root certificates list"
+    "composer/ca-bundle": "Provides regularly updated root certificates list"
   },
   "autoload": {
     "psr-4": {
@@ -48,6 +54,10 @@
     }
   },
   "scripts": {
+    "ci": [
+      "@phpcs",
+      "@test"
+    ],
     "test": "phpunit",
     "phpcs": "phpcs --standard=PSR2 src && phpcs --standard=PSR2 tests"
   }
