@@ -8,31 +8,20 @@ class UrlBuilder
 {
     const BASE_URL = 'https://www.fio.cz/ib_api/rest/';
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $token;
 
-    /**
-     * @param string $token
-     */
-    public function __construct($token)
+    public function __construct(string $token)
     {
         $this->setToken($token);
     }
 
-    /**
-     * @return string
-     */
-    public function getToken()
+    public function getToken(): string
     {
         return $this->token;
     }
 
-    /**
-     * @param string $token
-     */
-    public function setToken($token)
+    public function setToken(string $token)
     {
         if (!$token) {
             throw new MissingTokenException(
@@ -42,19 +31,30 @@ class UrlBuilder
         $this->token = $token;
     }
 
-    /**
-     * @param \DateTime $from
-     * @param \DateTime $to
-     *
-     * @return string
-     */
-    public function buildPeriodsUrl(\DateTime $from, \DateTime $to)
+    public function buildPeriodsUrl(\DateTimeInterface $from, \DateTimeInterface $to): string
     {
         return sprintf(
             self::BASE_URL.'periods/%s/%s/%s/transactions.json',
             $this->getToken(),
             $from->format('Y-m-d'),
             $to->format('Y-m-d')
+        );
+    }
+
+    public function buildLastUrl(): string
+    {
+        return sprintf(
+            self::BASE_URL . 'last/%s/transactions.json',
+            $this->getToken()
+        );
+    }
+
+    public function buildSetLastIdUrl(string $id): string
+    {
+        return sprintf(
+            self::BASE_URL . 'set-last-id/%s/%s/',
+            $this->getToken(),
+            $id
         );
     }
 }
